requirements.txt
absl-py==2.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
dill==0.4.0
filelock==3.16.1
flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.57.0
freetype-py==2.5.1
fsspec==2025.3.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
idna==3.10
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.7
lazy-loader==0.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyglet==2.1.6
PyOpenGL==3.1.0
pyparsing==3.1.4
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.21.0
scipy==1.10.1
six==1.17.0
smplx==0.1.28
sympy==1.13.3
tensorboard-data-server==0.7.2
tifffile==2023.7.10
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
trimesh==4.6.12
typing-extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.155
ultralytics-thop==2.0.14
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
watchfiles==0.21.0
websocket-client==1.7.0
Werkzeug==3.0.2
gunicorn
